# public/docs/swagger.json
{
  "info": {
    "title": "SusanooAPIServer",
    "version": "v0.1",
    "description": "Susanoo API Server"
  },
  "basePath": "/api/v1/",
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "description": "JWT authorization of an API",
      "name": "Authorization",
      "in": "header"
    }
  },
...
  "swagger": "2.0",
  "paths": {
...
    "/users/current": {
      "get": {
        "tags": [
          "Users"
        ],
        "name": "current",
        "summary": "get logging user details",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "User found and logged in successfully",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "mobileNumber": {
                      "type": "string"
                    },
                    "newEmail": {
                      "type": "string"
                    },
                    "location": {
                      "type": "string"
                    },
                    "isVerified": {
                      "type": "string"
                    },
                    "isConfirmedRegistration": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "patch": {
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update User info.",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "bio": {
                  "type": "string"
                },
                "countryId": {
                  "type": "string"
                },
                "facebookHandle": {
                  "type": "string"
                },
                "twitterHandle": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user updated successfully."
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
...
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "create new user",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "password": {
                  "type": "string",
                  "format": "password"
                },
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "mobileNumber": {
                  "type": "string",
                  "format": "phone"
                },
                "countryId": {
                  "type": "integer",
                  "format": "number"
                },
                "mobileCountryId": {
                  "type": "integer",
                  "format": "number"
                }
              }
            },
            "required": [
              "email_or_mobile_number",
              "password",
              "fingerprint"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "User found and logged in successfully.",
            "content": null,
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "mobileNumber": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "countryId": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "duplicate data"
          }
        }
      }
    },
...
    "/users/reset-password": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Reset password.",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "resetPasswordToken": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reset password process was successfully applie",
            "content": null,
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (VALIDATION_ERROR)"
          },
          "401": {
            "description": "Invalid signature (TOKEN_VERIFY_ERROR)"
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
...
    "/users/current/password": {
      "put": {
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Change User Password.",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "oldPassword": {
                  "type": "string"
                },
                "newPassword": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password changed.",
            "content": null,
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
...
    "/users/current/profile-image": {
      "post": {
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Upload user profile",
        "produces": [
          "multipart/form-data"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file",
            "description": "The file to upload."
          }
        ],
        "responses": {
          "200": {
            "description": "attachment add successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "duplicate data"
          }
        }
      },
      "delete": {
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Users"
        ],
        "summary": "Remove User Profile image",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "profile image removed"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
...
  },
...
}